+++ src/cjs/Circle.js
@@ -217,15 +217,13 @@ var Circle = /** @class */ (function () {
     };
     //--- BEGIN --- Implement interface `Intersectable`
     /**
-     * Get all line intersections with this polygon.
+     * Get all line intersections with this circle.
      *
      * This method returns all intersections (as vertices) with this shape. The returned array of vertices is in no specific order.
      *
-     * See demo `47-closest-vector-projection-on-polygon` for how it works.
-     *
      * @param {VertTuple} line - The line to find intersections with.
      * @param {boolean} inVectorBoundsOnly - If set to true only intersecion points on the passed vector are returned (located strictly between start and end vertex).
-     * @returns {Array<Vertex>} - An array of all intersections within the polygon bounds.
+     * @returns {Array<Vertex>} - An array of all intersections with the circle outline.
      */
     Circle.prototype.lineIntersections = function (line, inVectorBoundsOnly) {
         if (inVectorBoundsOnly === void 0) { inVectorBoundsOnly = false; }
@@ -248,7 +246,7 @@ var Circle = /** @class */ (function () {
      * This method returns all intersection tangents (as vectors) with this shape. The returned array of vectors is in no specific order.
      *
      * @param line
-     * @param inVectorBoundsOnly
+     * @param lineIntersectionTangents
      * @returns
      */
     Circle.prototype.lineIntersectionTangents = function (line, inVectorBoundsOnly) {