+++ src/ts/Circle.ts
@@ -267,15 +267,13 @@ export class Circle implements Intersectable, SVGSerializable {
 
   //--- BEGIN --- Implement interface `Intersectable`
   /**
-   * Get all line intersections with this polygon.
+   * Get all line intersections with this circle.
    *
    * This method returns all intersections (as vertices) with this shape. The returned array of vertices is in no specific order.
    *
-   * See demo `47-closest-vector-projection-on-polygon` for how it works.
-   *
    * @param {VertTuple} line - The line to find intersections with.
    * @param {boolean} inVectorBoundsOnly - If set to true only intersecion points on the passed vector are returned (located strictly between start and end vertex).
-   * @returns {Array<Vertex>} - An array of all intersections within the polygon bounds.
+   * @returns {Array<Vertex>} - An array of all intersections with the circle outline.
    */
   lineIntersections(line: VertTuple<any>, inVectorBoundsOnly: boolean = false): Array<Vertex> {
     // Find the intersections of all lines inside the edge bounds
@@ -297,7 +295,7 @@ export class Circle implements Intersectable, SVGSerializable {
    * This method returns all intersection tangents (as vectors) with this shape. The returned array of vectors is in no specific order.
    *
    * @param line
-   * @param inVectorBoundsOnly
+   * @param lineIntersectionTangents
    * @returns
    */
   lineIntersectionTangents(line: VertTuple<any>, inVectorBoundsOnly: boolean = false): Array<Vector> {